LayoutTests/http/tests/security/view-source-no-javascript-url-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "http". Protocols must match.

This test passes if it does not alert FAIL.
